Add a meson build for using as a wrap
The meson build system supports a concept of wraps, a dependency
management mechanism by which the build system can pull down
dependencies that are unmet by the system and build them itself. This
can solve version problems and the problem of tracking submodules.

This meson build is specifically only for using as a wrap, it builds the
three cpp files into a convenience library, and then exposes a meson
dependency object that will ink against this convenience library and
sets the proper include path.

This allows a project using meson as a build system to drop a simple ini
file (provided in the examples folder) into their project, add a couple
of simple calls to their build system, and get the latest git build of
imgui in their project. If/when a stable release is available an ini
file for the stable release could be added to build the stable release.